CNF-12680:CNF-13488: Add performance profile status for hypershift support (#1089)
* hypershift: declare consts for status implementation

Declare a set of consts that will be used for
the status report of `PerformanceProfile` when running
on Hypershift platform.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao: hypershift: refactor functions to reduce code duplication

GetTunedProfilesMessage, CalculateUpdatedStatus and ConfigMapMeta will be reused for hypershift performance profile status implementation.

Co-authored-by: Ronny Baturov <rbaturov@redhat.com>
Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* nto: export the logic for unique tuned name

On hypershift, `Tuned` object is being mirrored to the hosted cluster.
NTO makes `Tuned` names unique in case that Tuned is duplicated across `NodePool`s,
for example, if one `ConfigMap` is referenced in multiple `NodePool`s.

On OCP we used to list `Tuned` under the status of the
`PerformanceProfile`.

In order to achieve same behavior for hypershift,
we're capturing the logic, that grants unique name to `Tuned` objects,
into a function and export it.

On later commit, we'll use it to capture the name under the status.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao: hypershift: implement status writer interface

Wrting the \`PerformanceProfile\` status into a separate \`ConfigMap\`
that will be consumed later on by the Hypershift controller
and will be reported under the \`NodePool.status\`."

Co-authored-by: Ronny Baturov <rbaturov@redhat.com>
Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift:client: embed status in `PerformanceProfile`

On hypershift, PerformanceProfile's status is reported by a separate ConfigMap,
but for testing purposes we want to have same API as standard OCP.
IOW, we want to be able to access the status from the PerformanceProfile itself as follow:
`profile.status`

In order to achieve that, when user performs Get/List operation for PerformanceProfile,
the ControlPlane client will look for the ConfigMap that holds
the status (if it exists), extracts the status from it and embed it in the PerformanceProfile.

NOTE:
At this point in time the `Get()` and `List()` calls are done only
during the tests.
So this extra work of merging 2 ConfigMaps into the same `PerformanceProfile`
is quite expensive but is done only on testing anyway.
In the future, if we'll have to call `Get()`/`List()` with the `PerformanceProfile`
as part of the production code, we'll probably evaluate this approach
again.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* status: protection against tampering

The names of the tuned and runtime class are not pretected from user tampering.
We should validate the expected names against the existing ones and
reconcile if needed.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e: check for errors

Assert in case of error.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e:hypershift: use the new client

Replace the obsolete Client with ControlPlane/DataPlane client.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* client:dataplane: more granual access for objects

So far, we had objects that are accessible on the control plane (management cluster)
OR objects that are accessible on the data plane (hosted cluster).

Tuned is a special object which gets created on the control plane but then
mirrored into the data plane.
Sometime we want to read Tuned from the data plane, but we still don't want
to allow its creation on the data plane directly.

Hence, we provides a different checks for Read operation and Write operarion
as part of the data plane implementation.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e: add openshift label

This test should be running only on Openshift since MCP
not exist on Hypershift.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift: client: add `Delete` call

The `Delete` call deletes the `ConfigMap` that holds an encoded data of the object that was given for deletion.
This is needed for keeping backward compatability with API we're using for the
testing.
ref: https://github.com/openshift/cluster-node-tuning-operator/blob/10bb4c6a1388c91ee4184945e6103c82608eef27/test/e2e/performanceprofile/functests/3_performance_status/status.go#L153

NOTE:
This flow is not being used in production code ATM.
In case it does in the future, we might reassess this implementation if needed.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift: add the yamlSerializer as struct parameter

Use the same yaml serializer accross the methods of `ControlPlaneClientImpl`.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* controller: add tuned.profile predicates

We need to watch for tuned profile changes and update
the `PerformanceProfile` accordingly

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e: helper functions to attach/deattach objects to `NodePool`

On Hypershift, one should associate a tuning objects
 (`Tuned`/`performanceProfile`) with the NodePool
in order to apply configuration on the hosted cluster.

This commit provides helper functions for that
which shall be used while writing the e2e testing.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e:tuned: get tuned object name

Tuned object name is different when NTO runs on Hypershift.
This commit provides an helper function that returns
the correct name based on the platform (OCP/HCP)

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* vendor: gcustom

Needed for easier writing of custom mtacher for one of the status_suite\'s test.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e:status: sort imports

sort the imports according to the known convention applies on this repo.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e:status: remove cleaning function

A bit convoluted and not necessarily needed. we'll use `defer` in a follow-up commit instead.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e:status: adjust suite for hypershift

Make the necessary changes in order to be able to run this
suite over HCP platform.

The rationle some of the changes is capture in the code itself
for future references.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>
Co-authored-by: Talor Itzhak <titzhak@redhat.com>