NO-JIRA:e2e:hypershift: remove/rename deprecated env variables (#1121)
* e2e:hypershift: remove HYPERSHIFT_HOSTED_CLUSTER_KUBECONFIG env

This env is deprecated since KUBECONFIG is already pointing to the hosted-cluster kubeconfig.

Let's avoid confusion and redundancy and get rid of it.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e:hypershift: replace `HYPERSHIFT_MANAGEMENT_CLUSTER_NAMESPACE`

The name `HYPERSHIFT_HOSTED_CONTROL_PLANE_NAMESPACE` describes the purpose
of this Namespace in a more accurate way, since this namespace refers to
the namespace where all the hosted controlplane componenets reside.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e:hypershift: delete unused const

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>