ppc: getClusterData -> makeNodesHandlerForMCP
rename and refactor `getClusterData` function

clusterData contains the information about all the MCPs and their correlated nodes.

But the created profile should match a specific MCP so let's
simplifying the function to return a slice of nodes handlers that are matching only the mcp which correlate with the given mcp name.

 Signed-off-by: Talor Itzhak <titzhak@redhat.com>