ppc: info subcommand
When the info flag specified, ppc executes a completely different flow
and requires subset of parameters and validation.

In order to make the flow more consistent and readable
we introduce info as a separate subcommand.

NOTE: this change requires a tiny change in the command line argument API of PPC.
As disscussed offline, API guarantees do not include the tools command line.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>