ppc: delete getDataFromFlags
The filling of `ProfileCreatorArgs` is already done in the initFlags() function
so it's almost redundant but some flags validation that it's doing.

Let's move the flags validation into a dedicated function and delete
\`getDataFromFlags\`.
Let's move it into a dedicated function.

We also added the default value for `userLevelNetworking` to make the
tests happy. but no inteded changes in the actual behavior of the tool
itself.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>