ppc:hypershift: detect only once
Detecting whether the profile should be created for Hypershift is quite expensive operation.
It involves going over directories, opening file, and decoding it into object.
in addition it might return an error which make the API less convenient.

Instead we suggest to detect only once in the \`PersistentPreRunE\`
function and pass the value to all the relevant function during the ppc
runtime.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>