CNF-13391:ppc: adjust for Hypershift (#1124)
* ppc: decouple profile configuration from creation

Separate the configuration of the profile using the `profileData`

from the actual writing of the profile to the file-system.

The rationale is to make room for encapsulating the profile in a `ConfigMap` when creating it for a Hypershift platform.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: move flags initialization and info mode into functions

Improves code readability and devides this huge monolith
into separate units (functions) were each unit is resposible
for a single tasks.

In the nutshell this is just basic best-practice coding.
No intended changes in functionality.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: sort imports

std

3rd party

k8s

ocp

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: must gather validation to function

must gather's path is mandatory in every operation of ppc.

let's validate once at the beginning of execution, and then it can be used freely all over the code base.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: getClusterData -> makeNodesHandlerForMCP

rename and refactor `getClusterData` function

clusterData contains the information about all the MCPs and their correlated nodes.

But the created profile should match a specific MCP so let's
simplifying the function to return a slice of nodes handlers that are matching only the mcp which correlate with the given mcp name.

 Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: getProfileData -> makeProfileDataFrom

revamp and rename`getProfileData` function

We would like to use this function for Hypershift in a future commit.

On Hypershift there's no MCP concept, so let's extract the MCP logic from the function.

 Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: delete getDataFromFlags

The filling of `ProfileCreatorArgs` is already done in the initFlags() function
so it's almost redundant but some flags validation that it's doing.

Let's move the flags validation into a dedicated function and delete
\`getDataFromFlags\`.
Let's move it into a dedicated function.

We also added the default value for `userLevelNetworking` to make the
tests happy. but no inteded changes in the actual behavior of the tool
itself.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: info subcommand

When the info flag specified, ppc executes a completely different flow
and requires subset of parameters and validation.

In order to make the flow more consistent and readable
we introduce info as a separate subcommand.

NOTE: this change requires a tiny change in the command line argument API of PPC.
As disscussed offline, API guarantees do not include the tools command line.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: move validations to pre run functions

Validations should be performed in `PreRunE` and `PersistentPreRunE` functions
so the arguments could be used freely in the the `RunE` function.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: add node-pool-name flag

This flag is similar to mcp-name when must-gather is taken from a hosted-cluster which is running on
Hypershift platform.

It used to determine the subset of nodes from the cluster which the
profile will be created for.

- If mcp-name is provided we'll treat it as the node-pool-name.
- If both mcp-name and node-pool-name provided, an error will be raised.
- If none provided, an error will be raised

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: detect hypershift platform

Provides a way for ppc to detect if profile should be generated for a hypershift platform.

 Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: only list nodes associated with pool

Provides a unified way to list only the nodes which are matching the pool name provided by the user.

The function performs different logic based on the platform.

 Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: selectors handling

The  `nodeSelector` and `mcpSelector` values are diverged between hypershift and non-hypershift.
This commit handles the selector logic based on the platform.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: make profile for hypershift

return profile encapsulated in `ConfigMap` in case that
the profile is intended for Hypershift.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: adjust info command for hypershift

Adjust info command to detect whether it's running on Hypershift
and print a list of all node pools with there corresponding nodes.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: wrap error

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc:hypershift: detect only once

Detecting whether the profile should be created for Hypershift is quite expensive operation.
It involves going over directories, opening file, and decoding it into object.
in addition it might return an error which make the API less convenient.

Instead we suggest to detect only once in the \`PersistentPreRunE\`
function and pass the value to all the relevant function during the ppc
runtime.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: init flags in a receiver function

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc: test-unit data

Add test data for must-gather when running on hosted-cluster

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* ppc:tests: remove ioutil

Deprecated. Using `os` instead.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>