E2E: Add test to verify cpuset.cpus.exclusive is writeable (#1127)
Automates OCPBUGS-34812: cgroupsv2: failed to write on cpuset.cpus.exclusive

To reproduce the bug, we need to create and delete deployment (deploying
guaranteed pods with cpu load balancing annotation) in quick
succession, so that we do not fully wait for the cleanup causing the pod
about to be deleted to still have access to exclusive cpus causing the new pod
to fail because cpuset.cpus.exclusive is not yet freed.

As the pre-start hook fails to write to cpuset.cpus.exclusive file in the pods cgroup
pod goes to RunContainerError state.

This automation PR creates and deletes deployment in loop to reproduce
the issue and checks if the pods fails with Runtime error with message
"failed to run pre-start hook for container"

Signed-off-by: Niranjan M.R <mrniranjan@redhat.com>
Co-authored-by: Niranjan M.R <mrniranjan@redhat.com>