Add a script to test internal TuneD FDP releases (#1167)
This change adds a script to test internal TuneD FDP releases
hack/test-tuned-fdp.sh.  The script builds on top of
hack/deploy-custom-nto.sh script, but it makes use of Dockerfile.rhel9
versus the default upstream Dockerfile.

For detail on how to use it, refer to "Example invocation" section within
the script.  At a minimum, you'll typically want to supply your quay.io
username and baseurl to the TuneD FDP repository.

Co-authored-by: Jiri Mencak <jmencak@users.noreply.github.com>