NO-JIRA: hypershift:handler: add KubeletConfig unique label (#1169)
* handler: `EncapsulateObjInConfigMap` accepts multiple labels

We need to add a special label to KubeletConfig ConfigMap objects
so lets expand `EncapsulateObjInConfigMap` to accepts multiple labels
instead of of just one.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift:handler: add KubeletConfig unique label

According to https://github.com/openshift/enhancements/pull/1671
HyperShift operator will mirror KubeletConfig's ConfigMap to the hosted clusters.

In order to identify the desired ConfigMaps in efficient way,
HyperShift operator lists the ConfigMaps by this unique label.

KubeletConfig's ConfigMap which is generated by NTO should be mirror as well
hence the reason for adding the label.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>