E2E: wait for ovs services affinity to reset after deployment deletion (#1160)
In our test we create a deployment to deploy 2 Guranteed pods,
this affects the affinity  of ovs services , which gets changed
depending on the cpus used by Guaranteed pods,

once the deployment is delete, it takes time for ovs services to
return to the previous affinity. In this PR we
wait for the ovs processes to return to their original affinity
before executing the next test.

Signed-off-by: Niranjan M.R <mrniranjan@redhat.com>
Co-authored-by: Niranjan M.R <mrniranjan@redhat.com>