e2e: change tuned command exec handling
Since we change to `TTY: false` TuneD returns the following error:
"Cannot talk to TuneD daemon via DBus. Is TuneD daemon running?"
but yet provides the correct info from stdout.

Let's redirect the error to /dev/null so it won't fail the test.
Because ignoring errors is usually bad practice,
the test checks to see that it gets the expected tuned profile
from the command's output.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>