e2e: select RT Worker Nodes Label based on platform
In OCP, we use the worker-cnf label in the nodeSelector to target RT worker nodes when deploying an MCP with a matching nodeSelector. However, in Hypershift, all worker nodes within a nodepool (with a performance profile deployed) are considered RT worker nodes, so the default worker node label is sufficient for targeting an RT worker node (assuming we have one nodepool).

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>