e2e: modify code to be platform agnostic
As the RoleWorkerCNF (worker-cnf label) is specific to OCP and not used in Hypershift (which use the default worker label), we will update the code to print the dynamic value of NodeSelectorLabels instead.

Note: There are additional references to RoleWorkerCNF in the tests, but all of these are tests that run exclusively on OCP (as they are marked with the OpenShift label).

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>