e2e: remove PostUpdateSync and EnforceNodeLabels func
Previously, for e2e tests on Hypershift, we required labeling RT worker nodes with the worker-cnf label for one specific reason: the failure of test CNF-14628, where the pod couldn’t be scheduled on a node without the worker-cnf label.
Due to time constraints to merge the hypershift 1_performance suite adaptation, we introduced a workaround for this:  label the worker nodes with the worker-cnf label, using the  PostUpdateSync and EnforceNodeLabels functions. (Relabeling after each tuning application is necessary because the node loses its label once the tuning is applied).

We recently discovered that the test pod inherited the worker-cnf label after admission because the runtime class it referenced used the worker-cnf nodeSelector, which was originally defined in the performance profile. In the first commit of this PR, we modified this logic to create a test profile with the default worker label instead of worker-cnf, which should resolve the issue. As a result, the EnforceNodeLabels and PostUpdateSync functions are now redundant and will be removed.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>