NO-JIRA: Default to username and timestamp for custom NTO push
This improves the hack/deploy-custom-nto.sh script to not require
the ORG variable all the time. It defaults to username which users
commonly use for their quay.io as well.

The second change makes sure the newly built image has an unique
tag formed of short git hash and build timestamp. This helps
when systems are configured to use cached images. Each tag is
unique and so the cluster will always pull the new build.