OCPBUGS-43666: Fix kernel arguments ordering on Intel (#1197)
* Fix kernel arguments ordering on Intel

An upgrade from previous version causes one extra reboot
due to differently ordered kernel arguments. This is
a side effect of platform specific tuned profile split
we merged in https://github.com/openshift/cluster-node-tuning-operator/pull/1083

This fix updates the Intel specific tuned profile to
follow the same ordering that was used in the past.

It does so by exploiting a specific tuned behavior
of the bootloader plugin. It orders the kernel argument
cmdline_suffix keys based on the order of first appearance.
Any additional appearance just changes the value, but
not the ordering.

The change is only needed for Intel, because we have
never supported other platforms before and so upgrade
is not an issue.

* Sync rendered manifests