[release-4.17] OCPBUGS-43566: CI: unblock (#1188)
* e2e: change Tty to false

During latest changes done on OCP side,
we were start noticing metadate in the beginning
of the output returend from the exec call.

changing TTY to false seems to fix the issue.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e: check error returned from function call

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e: skip cpu_manager_state file state verification

Skip the test until https://issues.redhat.com/browse/OCPBUGS-43280
gets resolved.

This is done in order to unblock the CI for now.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e: replace split "\r\n" with "\n"

due to the recent changes in the shell configuration
(which we don't understand the root cause for)

we need to adjust the split function to make it work.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e: change tuned command exec handling

Since we change to `TTY: false` TuneD returns the following error:
"Cannot talk to TuneD daemon via DBus. Is TuneD daemon running?"
but yet provides the correct info from stdout.

Let's redirect the error to /dev/null so it won't fail the test.
Because ignoring errors is usually bad practice,
the test checks to see that it gets the expected tuned profile
from the command's output.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>
Co-authored-by: Talor Itzhak <titzhak@redhat.com>