nodeInspector: do not reuse namespace
Using the global namespace object leads to an error
when trying to recreate the namespace againg.

This is becuase when the first `Create` call return
the object is fullfiled with some metadata from the API server
like `resourceVersion`.

This in turn leads to error when trying to reuse the object for creation
becuase the API doesn't expect the `resourceVersion` to be set.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>