OCPBUGS-38900: Drop sched_migration_cost_ns setting (#1203)
When using a PerformanceProfile, a TuneD scheduler was setting
sched_migration_cost_ns.  This setting was previously used by
throughput-performance and latency-performance profiles.  However, it
was dropped from these profiles based on performance evaluation for
RHEL-9 as discussed in rhbz#1957829.

Setting sched_migration_cost_ns causes issues when SecureBoot/kernel_lockdown
is in effect.  Do not set this tunable when using PerformanceProfiles.

Resolves: OCPBUGS-38900

Co-authored-by: Jiri Mencak <jmencak@users.noreply.github.com>