OCPBUGS-44283: right-hand-side profile_dirs take precedence
This is a manual backport of #1061.

The new TuneD option profile_dirs defines custom TuneD directories. Fix
the order of the directories so that the system TuneD directory
(/usr/lib/tuned) is the first one (searched last), and the custom TuneD
directory (/var/lib/ocp-tuned/profiles) is the last one (searched
first).

Resolves: OCPBUGS-44283