refactor container runtime selection logic for 4.18 compatibility
Previously, to select the container runtime for high-performance class, we checked for an existing ContainerRuntimeConfig. If none was found, the default runtime was used. However, starting with OCP 4.18, the default runtime switches to crun, meaning that systems upgrading from OCP 4.17 to 4.18 would experience an unintended runtime switch under the hood.
To prevent this, the MCO team introduced a patch in OCP 4.17 to create a default machine config for all MCP's, explicitly setting runc as the runtime. As a result, we need to adjust our logic to identify this MachineConfig when no ContainerRuntimeConfig is found.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>