Manage TuneD profiles with the same name and different content
It is possible to create Tuned CRs with TuneD profiles of the same name.
This is problematic when the duplicate TuneD profiles have different
contents.  This can cause periodic switches between the duplicate TuneD
profiles and will lead to TuneD restarts.

This change makes the operator control loop stop to avoid switching
between the conflicting TuneD profiles of the same name and warns about
this misconfiguration via:
  * Tuned CR status and prometheus metric (nto_invalid_tuned_exist_info)
  * NTOInvalidTunedExist alert
  * ERROR message issued in the operator logs

Other changes:
  * Switch the ProfileStatusCondition condition type to
    StatusCondition.  This is cosmetic only, all fields are preserved.
  * Obsolete pod label-based matching.  Trigger an alert
    (NTOPodLabelsUsed) for this.

Fixes: OCPBUGS-44559.