Remove detecting  runtime from containerRuntimeConfig
A new patch has been introduced for CRI-O, allowing the container runtime for the high-performance runtime class to be inherited from the system instead of being detected and explicitly set. cri-o/cri-o#8753

As a result, the controller no longer needs to identify the container runtime from the containerRuntimeConfig for setting it in the 99-runtimes.conf snippet.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>