OCPBUGS-44372: PPC: skip comparing ProcessorCore.Index between NUMA cores  (#1213)
* PPC: skip comparing ProcessorCore.Index between NUMA cores

ProcessorCore.Index indicates the zero-based index of the core in the
Cores slice. While core might be shown in a different order, they can still
be equivalent. See: jaypipes/ghw#346.

Adjust the equality check to skip this field to fix this:

```
  Error: targeted nodes differ: nodes host1.development.lab and host2.development.lab have different topology: the CPU cores differ: processor core #20 (2 threads), logical processors [2 66] vs processor core #20 (2 threads), logical processors [2 66]
```

And add a unit test to cover this scenario.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>

* PPC: unit: rename test variables to avoid misusing

Rename test variables and add clarifying comments to avoid misusing them
while writing tests.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>

---------

Signed-off-by: Shereen Haj <shajmakh@redhat.com>