e2e: ContainerRuntimeConfig test logic changes
When we have no ContainerRuntimeConfig in the system, the high-performance class should use the runtime that is currently being used in the system. This should be crun for 4.18 systems, or runc if upgraded from older version and not explicitly switched to crun.
Since the selected runtime is not directly exposed, we target the tuned pod to determine which runtime is handling it, using this as an indicator for the node's default runtime.
Moreover, instead of checking the container runtime for the high-performance pods from the conf file (which no longer is possible), we will create a test pod and check the container runtime used in practice.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>