CNF-15505: Remove container runtime selection (#1219)
* Remove detecting  runtime from containerRuntimeConfig

A new patch has been introduced for CRI-O, allowing the container runtime for the high-performance runtime class to be inherited from the system instead of being detected and explicitly set. cri-o/cri-o#8753

As a result, the controller no longer needs to identify the container runtime from the containerRuntimeConfig for setting it in the 99-runtimes.conf snippet.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* Removed redundant crio configuration

This snippet necessity has been questioned in the first time added, and perhaps wasn't needed to be added.
Moreover, we are now inheriting crio configuration .
The tests in the latter commit can verify this is redundant.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* controller: remove ctrcfg from being watched

Previously, we monitored the containerRuntimeConfig for changes to update the CRI-O configuration for the high-performance runtime class. Since this is no longer necessary, we should remove this logic.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* render: remove containerRuntimeConfig object

This commit removes rendering for the containerRuntimeConfig.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* e2e: detect runtime used by pod from node

The GetContainerRuntimeTypeFor function has been updated to provide more accurate output regarding the runtime class used on the node. Previously, it relied on reading the runtime path from the CRI-O configuration file, but this is no longer viable, as the runtime path is no longer written to the file. The new approach directly checks the container process on the node using the ps tool to determine the runtime in use.
The new command outputs the conmon process of the container and greps the container runtime from the -r argument using regex.

For example:
/bin/ps aux | grep 'tuned-ncp4j'
root        6317  0.0  0.0   8312  1600 ?        Ss   12:25   0:00 /usr/bin/conmon -b /run/containers/storage/overlay-containers/f492c7c342b71.../userdata/conmon-pidfile -p /run/containers/storage/overlay-containers/f492c7c342b71e19e.../userdata/pidfile --persist-dir /var/lib/containers/storage/overlay-containers/f492c7c342b7..../userdata -r /usr/bin/crun --runtime-arg --root=/run/crun --socket-dir-path /var/run/crio --syslog -u f492c7c3... -s
root       90775  0.0  0.0   3332  1600 ?        S+   13:57   0:00 grep tuned-ncp4j

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* e2e: ContainerRuntimeConfig test logic changes

When we have no ContainerRuntimeConfig in the system, the high-performance class should use the runtime that is currently being used in the system. This should be crun for 4.18 systems, or runc if upgraded from older version and not explicitly switched to crun.
Since the selected runtime is not directly exposed, we target the tuned pod to determine which runtime is handling it, using this as an indicator for the node's default runtime.
Moreover, instead of checking the container runtime for the high-performance pods from the conf file (which no longer is possible), we will create a test pod and check the container runtime used in practice.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

---------

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>