PPC: unit: add test with different siblings on same core ID
So far PPC checked that compared sockets are identical in terms of Core ID,
threads' number and logical processors list. The scope and importance of the Core ID is
mainly within the single socket (numa/dei) and by tracking it we can
know the siblings list for a specific core ID.

Add missing negative unit test for `ensureSameTopology()` where it
contructs 2 compute nodes that are completely identical, while on NUMA
level there are cores with different siblings list. This should be
reporting an error but it will fail because the verification logic is
not checked but will follow next.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>