PPC: verify core with the same ID always has same sibling list
Or in other words, verify that siblings CPUs are associated with same
core id on NUMA level.

Core ID (from `/sys/devices/system/node/nodeX/cpuY/topology/core_id`) is
important especially for the specific NUMA, and CPUs siblings are
featured by that they have same core id. Although this rarely happens in
real computers, PPC tool depends on the data reported by the
must-gather, so we need to verify that what's received is still
preserving the accepted configuration for generating a
performance-profile.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>