cmd: PPC: support tolerating heterogeneous core IDs
**Problem: In a system with a specific number of cores per socket
the host gives each core a number. So far PPC would generate
performance profile only after verifying that the nodes pointed to by
the specified node pool are all of the same hardware and topology. That
is because the performance profile will apply the same configuration on all
these nodes and it is most important that they have same structure, CPUs
distribution across NUMAs and CPUs availability. This by default
includes having same core IDs for each CPU in the same NUMA cells for
all the compute nodes. for instance:

Worker-0 has NUMA-0 on which the CPUs siblings [2,66] is coupled in
one core numbered 18`.
All other workers should have this info true for them, otherwise the
tool would fail.

**Suggested solution:**
With further investigation, it was learned that core numbering can have
different schemes even with a system from the same vendor. The numbering
the pattern depends on the settings of the hardware, the software and the
firmware (BIOS).While core IDs may vary nodes can still be considered
having same NUMA topology taking into account that core scope is on the
single NUMA. However, core IDs can be important in optimizing the
system's performance and managing isolation of tasks, meaning if the
performance of worker-0 is not comparable to that of worker-1 before
having performance-profile applied on both, having an improved
performance on both after applying PP is not guaranteed.

In this PR, we introduce an experimental flag to be added to the PPC
tool called `tolerate-core-ids-diff` and is `false` by default; if set
to true then the check to ensure same topologies between nodes will
ignore having different core IDs for the same CPUs in the same NUMA
cells.

Disclaimer: We support this option to unblock business matters and is recommended
to use it with caution only if the default value is not providing the
desired results; and only if the user is willing to take on the
responsibility of troubleshooting the settings on their environment which not all
are reflected in the must-gather collection.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>